Refactor cyral_repository_access_gateway
