+++ cyral/internal/repository/accessgateway/constants.go
@@ -1,0 +1,5 @@
+package accessgateway
+
+const (
+	resourceName = "cyral_repository_access_gateway"
+)
