+++ cyral/provider/schema_loader.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/slack"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/teams"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
+	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessgateway"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/datamap"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/useraccount"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/samlcertificate"
@@ -15,6 +16,7 @@ import (
 
 func packagesSchemas() []core.PackageSchema {
 	v := []core.PackageSchema{
+		accessgateway.PackageSchema(),
 		datalabel.PackageSchema(),
 		datamap.PackageSchema(),
 		hcvault.PackageSchema(),
